01:48  warehouse need to be corrected

+++ ProjectBackend/src/main/java/org/isdb/ProjectBackend/controller/login/tCon/CartController.java
@@ -26,37 +26,37 @@ public class CartController {
 	@Autowired
 	private CartService cartService;
 
-	// Get all carts
+	
 	@GetMapping
 	public List<Cart> getAllCarts() {
 		return cartService.getAllCarts();
 	}
 
-	// Get a cart by ID
+	
 	@GetMapping("/{id}")
-	public ResponseEntity<Cart> getCartById(@PathVariable Integer id) {
+	public ResponseEntity<Cart> getCartById(@PathVariable Long id) {
 		Optional<Cart> cart = cartService.getCartById(id);
 		return cart.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
 	}
 
-	// Create a new cart
+	
 	@PostMapping
 	public ResponseEntity<Cart> createCart(@RequestBody Cart cart) {
 		Cart savedCart = cartService.saveCart(cart);
 		return ResponseEntity.status(HttpStatus.CREATED).body(savedCart);
 	}
 
-	// Update an existing cart
+	
 	@PutMapping("/{id}")
-	public ResponseEntity<Cart> updateCart(@PathVariable Integer id, @RequestBody Cart cartDetails) {
+	public ResponseEntity<Cart> updateCart(@PathVariable Long id, @RequestBody Cart cartDetails) {
 		Cart updatedCart = cartService.updateCart(id, cartDetails);
 		return updatedCart != null ? ResponseEntity.ok(updatedCart)
 				: ResponseEntity.status(HttpStatus.NOT_FOUND).build();
 	}
 
-	// Delete a cart
+
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deleteCart(@PathVariable Integer id) {
+	public ResponseEntity<Void> deleteCart(@PathVariable Long id) {
 		cartService.deleteCart(id);
 		return ResponseEntity.noContent().build();
 	}
