+++ ProjectBackend/src/main/java/org/isdb/ProjectBackend/controller/login/tCon/WarehouseController.java
@@ -30,7 +30,7 @@ public class WarehouseController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<WarehouseDTO> getById(@PathVariable Integer id) {
+	public ResponseEntity<WarehouseDTO> getById(@PathVariable Long id) {
 		return warehouseService.getWarehouseById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
 	}
 
@@ -40,13 +40,13 @@ public class WarehouseController {
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<WarehouseDTO> update(@PathVariable Integer id, @RequestBody WarehouseDTO dto) {
+	public ResponseEntity<WarehouseDTO> update(@PathVariable Long id, @RequestBody WarehouseDTO dto) {
 		WarehouseDTO updated = warehouseService.updateWarehouse(id, dto);
 		return (updated != null) ? ResponseEntity.ok(updated) : ResponseEntity.notFound().build();
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> delete(@PathVariable Integer id) {
+	public ResponseEntity<Void> delete(@PathVariable Long id) {
 		return warehouseService.deleteWarehouse(id) ? ResponseEntity.noContent().build()
 				: ResponseEntity.notFound().build();
 	}
