+++ ProjectBackend/src/main/java/org/isdb/ProjectBackend/repository/DiscountRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiscountRepository extends JpaRepository<Discount, Integer> {
+public interface DiscountRepository extends JpaRepository<Discount, Long> {
     // You can add custom query methods here if needed
 }
